Some additions such as query strings and refractoring.
Scripts are now given query strings as arguments in the form of 'key=value'.
Probably not the best way to do it. If you look at 'test-site/blog/index'
you should get an idea of how it works.

So scripts seem to be working.

Returning markdown works.

And all in 252 lines.